+++ packages/salesforcedx-vscode-replay-debugger/src/constants.ts
@@ -4,7 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
+export const CHECKPOINT = 'checkpoint';
+export const CHECKPOINT_INFO_EVENT = 'checkpointInfo';
 export const DEBUGGER_LAUNCH_TYPE = 'launch';
 export const DEBUGGER_TYPE = 'apex-replay';
 export const DEFAULT_INITIALIZE_TIMEOUT_MS = 10000;